feat: refactor ColorMode into ActionDumpFormat
